requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
scipy=1.1.0
flask
